backend: tidy governance processor
